Extract String RNG into Random, hashing into Hash class.
1. This is to prepare the String for a rename (probably `Text`)
   these features don't fit into a `Text` class anymore.
2. These features have more to do with security.
3. The String class has too many different scopes and should
   purely deal with String/Text manipulation.
4. Packages inside the security NS have one dependency less
   outside themselves.
