+++ security/Password.php
@@ -8,7 +8,8 @@
 
 namespace lithium\security;
 
-use lithium\util\String;
+use lithium\security\Hash;
+use lithium\security\Random;
 
 /**
  * `Password` utility class that makes use of PHP's `crypt()` function. Includes a
@@ -95,7 +96,7 @@ class Password {
 	 * @return boolean Returns a boolean indicating whether the password is correct.
 	 */
 	public static function check($password, $hash) {
-		return String::compare($hash, crypt($password, $hash));
+		return Hash::compare($hash, crypt($password, $hash));
 	}
 
 	/**
@@ -124,13 +125,13 @@ class Password {
 	 * the resulting salts are prefixed with information expected by PHP's
 	 * `crypt()`. To get an arbitrarily long, cryptographically strong salt
 	 * consisting in random sequences of alpha numeric characters, use
-	 * `lithium\util\String::random()` instead.
+	 * `lithium\security\Random::generate()` instead.
 	 *
 	 * @link http://php.net/function.crypt.php
 	 * @link http://www.postgresql.org/docs/9.0/static/pgcrypto.html
 	 * @see lithium\security\Password::hash()
 	 * @see lithium\security\Password::check()
-	 * @see lithium\util\String::random()
+	 * @see lithium\security\Random::generate()
 	 * @param string $type The hash type. Optional. Defaults to the best
 	 *        available option. Supported values, along with their maximum
 	 *        password lengths, include:
@@ -156,7 +157,7 @@ class Password {
 
 	/**
 	 * Generates a Blowfish salt for use in `lithium\security\Password::hash()`. _Note_: Does not
-	 * use the `'encode'` option of `String::random()` because it could result in 2 bits less of
+	 * use the `'encode'` option of `Random::generate()` because it could result in 2 bits less of
 	 * entropy depending on the last character.
 	 *
 	 * @param integer $count The base-2 logarithm of the iteration count.
@@ -170,7 +171,7 @@ class Password {
 		$base64 = './ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 		$i = 0;
 
-		$input = String::random(16);
+		$input = Random::generate(16);
 		$output = '';
 
 		do {
@@ -218,7 +219,7 @@ class Password {
 
 		$output = '_' . $base64[$count & 0x3f] . $base64[($count >> 6) & 0x3f];
 		$output .= $base64[($count >> 12) & 0x3f] . $base64[($count >> 18) & 0x3f];
-		$output .= String::random(3, array('encode' => String::ENCODE_BASE_64));
+		$output .= Random::generate(3, array('encode' => Random::ENCODE_BASE_64));
 
 		return $output;
 	}
@@ -229,7 +230,7 @@ class Password {
 	 * @return string The MD5 salt.
 	 */
 	protected static function _generateSaltMd5() {
-		return '$1$' . String::random(6, array('encode' => String::ENCODE_BASE_64));
+		return '$1$' . Random::generate(6, array('encode' => Random::ENCODE_BASE_64));
 	}
 }
